+++ internal/config/options.go
@@ -33,6 +33,7 @@ type options struct {
 	version                        version.Version
 	logger                         logr.Logger
 	autoInstrumentationDotNetImage string
+	autoInstrumentationGoImage     string
 	autoInstrumentationJavaImage   string
 	autoInstrumentationNodeJSImage string
 	autoInstrumentationPythonImage string
@@ -44,8 +45,8 @@ type options struct {
 	onOpenShiftRoutesChange        changeHandler
 	labelsFilter                   []string
 	openshiftRoutes                openshiftRoutesStore
+	hpaVersion                     hpaVersionStore
 	autoDetectFrequency            time.Duration
-	autoscalingVersion             autodetect.AutoscalingVersion
 }
 
 func WithAutoDetect(a autodetect.AutoDetect) Option {
@@ -88,6 +89,7 @@ func WithLogger(logger logr.Logger) Option {
 		o.logger = logger
 	}
 }
+
 func WithOnOpenShiftRoutesChangeCallback(f func() error) Option {
 	return func(o *options) {
 		if o.onOpenShiftRoutesChange == nil {
@@ -128,6 +130,12 @@ func WithAutoInstrumentationPythonImage(s string) Option {
 func WithAutoInstrumentationDotNetImage(s string) Option {
 	return func(o *options) {
 		o.autoInstrumentationDotNetImage = s
+	}
+}
+
+func WithAutoInstrumentationGoImage(s string) Option {
+	return func(o *options) {
+		o.autoInstrumentationGoImage = s
 	}
 }