+++ pkg/collector/reconcile/configmap.go
@@ -76,7 +76,7 @@ func desiredConfigMap(_ context.Context, params Params) corev1.ConfigMap {
 	}
 	config, err := ReplaceConfig(params.Instance)
 	if err != nil {
-		params.Log.V(2).Info("failed to update prometheus config to use sharded targets: ", err)
+		params.Log.V(2).Info("failed to update prometheus config to use sharded targets: ", "err", err)
 	}
 
 	return corev1.ConfigMap{
@@ -114,7 +114,8 @@ func desiredTAConfigMap(params Params) (corev1.ConfigMap, error) {
 		"app.kubernetes.io/managed-by": "opentelemetry-operator",
 		"app.kubernetes.io/component":  "opentelemetry-collector",
 	}
-	taConfig["config"] = promConfig
+	// We only take the "config" from the returned object, we don't need the "target_allocator" configuration here.
+	taConfig["config"] = promConfig["config"]
 	if len(params.Instance.Spec.TargetAllocator.AllocationStrategy) > 0 {
 		taConfig["allocation_strategy"] = params.Instance.Spec.TargetAllocator.AllocationStrategy
 	} else {