+++ pkg/featuregate/featuregate.go
@@ -21,13 +21,43 @@ import (
 )
 
 const (
-	featureGatesFlag = "feature-gates"
+	FeatureGatesFlag = "feature-gates"
+)
+
+var (
+	EnableDotnetAutoInstrumentationSupport = featuregate.GlobalRegistry().MustRegister(
+		"operator.autoinstrumentation.dotnet",
+		featuregate.StageBeta,
+		featuregate.WithRegisterDescription("controls whether the operator supports .NET auto-instrumentation"))
+	EnablePythonAutoInstrumentationSupport = featuregate.GlobalRegistry().MustRegister(
+		"operator.autoinstrumentation.python",
+		featuregate.StageBeta,
+		featuregate.WithRegisterDescription("controls whether the operator supports Python auto-instrumentation"))
+	EnableJavaAutoInstrumentationSupport = featuregate.GlobalRegistry().MustRegister(
+		"operator.autoinstrumentation.java",
+		featuregate.StageBeta,
+		featuregate.WithRegisterDescription("controls whether the operator supports Java auto-instrumentation"))
+	EnableNodeJSAutoInstrumentationSupport = featuregate.GlobalRegistry().MustRegister(
+		"operator.autoinstrumentation.nodejs",
+		featuregate.StageBeta,
+		featuregate.WithRegisterDescription("controls whether the operator supports NodeJS auto-instrumentation"))
+	EnableGoAutoInstrumentationSupport = featuregate.GlobalRegistry().MustRegister(
+		"operator.autoinstrumentation.go",
+		featuregate.StageAlpha,
+		featuregate.WithRegisterDescription("controls whether the operator supports Goland auto-instrumentation"))
+
+	// EnableTargetAllocatorRewrite is the feature gate that controls whether the collector's configuration should
+	// automatically be rewritten when the target allocator is enabled.
+	EnableTargetAllocatorRewrite = featuregate.GlobalRegistry().MustRegister(
+		"operator.collector.rewritetargetallocator",
+		featuregate.StageAlpha,
+		featuregate.WithRegisterDescription("controls whether the operator should configure the collector's targetAllocator configuration"))
 )
 
 // Flags creates a new FlagSet that represents the available featuregate flags using the supplied featuregate registry.
 func Flags(reg *featuregate.Registry) *flag.FlagSet {
 	flagSet := new(flag.FlagSet)
-	flagSet.Var(featuregate.NewFlag(reg), featureGatesFlag,
+	flagSet.Var(featuregate.NewFlag(reg), FeatureGatesFlag,
 		"Comma-delimited list of feature gate identifiers. Prefix with '-' to disable the feature. '+' or no prefix will enable the feature.")
 	return flagSet
 }